refactor: structure component previews
